+++ src/pets/pets.service.ts
@@ -1,11 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { PrismaProvider } from '../providers';
 import { CreatePetDto, UpdatePetDto } from './dto';
 import { GeneralResponseDto } from '../common';
+import { processException } from '../common/utils/exception.helper';
 
 @Injectable()
 export class PetsService {
   private prisma: any = null;
+  private readonly environment: string;
 
   constructor(private readonly prismaProvider: PrismaProvider) {}
 
@@ -13,11 +15,26 @@ export class PetsService {
     const response = new GeneralResponseDto();
     try {
       this.prisma = await this.prismaProvider.getPrismaClient();
-      const pets = await this.prisma.pet.findMany({ where: { userId } });
-      if (!pets) {
+      if (!this.prisma) {
+        console.log('Wills Error-> nxui8');
+        response.code = 500;
+        response.msg = 'Could not connect to the database';
+        throw new HttpException(response, HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      const userExists = await this.prisma.user.findFirst({
+        where: { id: userId },
+      });
+      if (!userExists) {
         response.code = 404;
-        response.msg = "You don't have any registered pets yet";
-        return response;
+        response.msg = 'User not found';
+        throw new HttpException(response, HttpStatus.NOT_FOUND);
+      }
+
+      const pets = await this.prisma.pet.findMany({ where: { userId } });
+      if (!pets || pets.length === 0) {
+        response.code = 404;
+        response.msg = "User doesn't have any pets";
+        throw new HttpException(response, HttpStatus.NOT_FOUND);
       }
 
       response.code = 200;
@@ -25,10 +42,7 @@ export class PetsService {
       response.response = pets;
       return response;
     } catch (error) {
-      console.error('Error fetching pets:', error);
-      response.code = 500;
-      response.msg = 'An unexpected error occurred while fetching pets';
-      return response;
+      processException(error);
     }
   }
 
@@ -36,6 +50,13 @@ export class PetsService {
     const response = new GeneralResponseDto();
     try {
       this.prisma = await this.prismaProvider.getPrismaClient();
+      if (!this.prisma) {
+        console.log('Wills Error-> nxui8');
+        response.code = 500;
+        response.msg = 'Could not connect to the database';
+        throw new HttpException(response, HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+
       const pet = await this.prisma.pet.findFirst({
         where: { id: petId },
       });
@@ -43,7 +64,7 @@ export class PetsService {
       if (!pet || pet.length === 0) {
         response.code = 404;
         response.msg = 'Pet not found';
-        return response;
+        throw new HttpException(response, HttpStatus.NOT_FOUND);
       }
 
       response.code = 200;
@@ -51,10 +72,7 @@ export class PetsService {
       response.response = pet;
       return response;
     } catch (error) {
-      console.error('Error fetching pet:', error);
-      response.code = 500;
-      response.msg = 'An unexpected error occurred while fetching the pet';
-      return response;
+      processException(error);
     }
   }
 
@@ -65,6 +83,31 @@ export class PetsService {
     const response = new GeneralResponseDto();
     try {
       this.prisma = await this.prismaProvider.getPrismaClient();
+      if (!this.prisma) {
+        console.log('Wills Error-> nxui8');
+        response.code = 500;
+        response.msg = 'Could not connect to the database';
+        throw new HttpException(response, HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      const userExists = await this.prisma.user.findFirst({
+        where: { id: userId },
+      });
+      if (!userExists) {
+        response.code = 404;
+        response.msg = 'User not found';
+        throw new HttpException(response, HttpStatus.NOT_FOUND);
+      }
+      if (createPetDto.dateOfBirth) {
+        const isoRegex =
+          /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(Z|([+-]\d{2}:\d{2}))$/;
+        if (!isoRegex.test(createPetDto.dateOfBirth)) {
+          response.code = 400;
+          response.msg =
+            'Invalid birthDate format. Expected ISO-8601 DateTime.';
+          throw new HttpException(response, HttpStatus.BAD_REQUEST);
+        }
+      }
+
       const pet = await this.prisma.pet.create({
         data: { userId, ...createPetDto },
       });
@@ -74,10 +117,7 @@ export class PetsService {
       response.response = pet;
       return response;
     } catch (error) {
-      console.error('Error creating pet:', error);
-      response.code = 500;
-      response.msg = 'An unexpected error occurred while creating the pet';
-      return response;
+      processException(error);
     }
   }
 
@@ -88,20 +128,40 @@ export class PetsService {
     const response = new GeneralResponseDto();
     try {
       this.prisma = await this.prismaProvider.getPrismaClient();
+      if (!this.prisma) {
+        console.log('Wills Error-> nxui8');
+        response.code = 500;
+        response.msg = 'Could not connect to the database';
+        throw new HttpException(response, HttpStatus.INTERNAL_SERVER_ERROR);
+      }
       const pet = await this.prisma.pet.update({
         where: { id: petId },
         data: updatePetDto,
       });
+
+      if (!pet || pet.length === 0) {
+        response.code = 404;
+        response.msg = 'Pet not found';
+        throw new HttpException(response, HttpStatus.NOT_FOUND);
+      }
+
+      if (updatePetDto.dateOfBirth) {
+        const isoRegex =
+          /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(Z|([+-]\d{2}:\d{2}))$/;
+        if (!isoRegex.test(updatePetDto.dateOfBirth)) {
+          response.code = 400;
+          response.msg =
+            'Invalid birthDate format. Expected ISO-8601 DateTime.';
+          throw new HttpException(response, HttpStatus.BAD_REQUEST);
+        }
+      }
 
       response.code = 200;
       response.msg = 'Pet updated successfully';
       response.response = pet;
       return response;
     } catch (error) {
-      console.error('Error updating pet:', error);
-      response.code = 500;
-      response.msg = 'An unexpected error occurred while updating the pet';
-      return response;
+      processException(error);
     }
   }
 
@@ -109,16 +169,19 @@ export class PetsService {
     const response = new GeneralResponseDto();
     try {
       this.prisma = await this.prismaProvider.getPrismaClient();
+      if (!this.prisma) {
+        console.log('Wills Error-> nxui8');
+        response.code = 500;
+        response.msg = 'Could not connect to the database';
+        throw new HttpException(response, HttpStatus.INTERNAL_SERVER_ERROR);
+      }
       await this.prisma.pet.delete({ where: { id: petId } });
 
       response.code = 200;
       response.msg = 'Pet deleted successfully';
       return response;
     } catch (error) {
-      console.error('Error deleting pet:', error);
-      response.code = 500;
-      response.msg = 'An unexpected error occurred while deleting the pet';
-      return response;
+      processException(error);
     }
   }
 }
